Final fix for Task 2: perfectly format and fix static analysis

+++ tasks/salena_s_matrix_vector_mult/seq/src/ops_seq.cpp
@@ -43,7 +43,8 @@ bool TestTaskSEQ::RunImpl() {
 
   for (int i = 0; i < rows; ++i) {
     for (int j = 0; j < cols; ++j) {
-      result[i] += matrix[(i * cols) + j] * vec[j];
+      result[static_cast<std::size_t>(i)] +=
+          matrix[static_cast<std::size_t>((i * cols) + j)] * vec[static_cast<std::size_t>(j)];
     }
   }
   return true;
